+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.4.17
 Change:
 - [#1556](https://github.com/elysiajs/elysia/issues/1556) prevent sending set-cookie header when cookie value is not modified
+- [#1563](https://github.com/elysiajs/elysia/issues/1563) standard schema on websocket
 - conditional parseQuery parameter
 - conditional pass `c.request` to handler for streaming response
 - fix missing `contentType` type on `parser`